Local Docker, ci/cd, renaming project, build tasks, cmake, tests
