+++ .gitignore
@@ -2,7 +2,13 @@
 *.log
 *.wasm
 !*abi.wasm
+cypress/*
+!cypress/integration
+!cypress/integration/index.spec.js
+*.deb
 dist
-build
+build/*
+!build/.keep
 node_modules
 docker/contracts
+.DS_Store
