+++ .gitpod.yml
@@ -1,20 +1,17 @@
-image: bradlhart/sample-return-values:v0.0.4
+image: eosio/return-values-example:v0.1.0
 
 ports:
 - port: 3000
   onOpen: ignore
-- port: 8000
-  onOpen: open-preview
 - port: 8080
-  onOpen: ignore
+  onOpen: open-preview
 - port: 8888
   onOpen: ignore
 - port: 9876
   onOpen: ignore
 
 tasks:
-- before: cd webapp
-  init: yarn
-  command: nginx -c $PWD/nginx.conf; npx webpack-dev-server
-- before: cleos wallet unlock --password </password
-  command: nodeos --config-dir ~/eosio/chain/config --data-dir ~/eosio/chain/data -e -p eosio --plugin eosio::chain_api_plugin --contracts-console
+- before: yarn
+  command: yarn run serveInGitPod
+- before: chmod 755 $PWD/docker/deploy_contracts.sh
+  command: $PWD/docker/deploy_contracts.sh "" IS_GITPOD
