+++ contracts/action_return_value/ricardian/action_return_value.contracts.md
@@ -1,0 +1,10 @@
+<h1 class="contract">sum</h1>
+
+---
+spec_version: "0.2.0"
+title: Sum
+summary: 'This action will add the two provided integers and return the result.'
+icon:
+---
+
+This action will add the two provided integers and return the result to demonstrate the action return values feature.
