+++ index.html
@@ -9,11 +9,11 @@
 <html>
     <head>
         <meta charset="UTF-8" />
-        <title>Sample Return Values Application</title>
+        <title>Return Values Example App</title>
     </head>
 
     <body>
-        <div id="sample"></div>
+        <div id="example"></div>
         <script src="./main.js"></script>
     </body>
