+++ package.json
@@ -1,0 +1,33 @@
+{
+  "private": true,
+  "description": "Example Website",
+  "main": "index.html",
+  "scripts": {
+    "serve": "webpack-dev-server --mode development",
+    "serveInGitPod": "webpack-dev-server --mode development",
+    "docker-run": "docker-compose down -v && docker-compose up",
+    "docker-persistent": "docker-compose up",
+    "lint": "eslint --ext .js,.jsx,.ts,.tsx src"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@blockone/eslint-config-blockone": "^3.0.0",
+    "@types/react": "^16.14.2",
+    "@types/react-dom": "^16.9.10",
+    "css-loader": "^4.3.0",
+    "eslint": "^6.8.0",
+    "source-map-loader": "^1.1.3",
+    "style-loader": "^1.3.0",
+    "ts-loader": "^8.0.11",
+    "typescript": "^3.9.7",
+    "webpack": "^4.44.2",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
+  },
+  "dependencies": {
+    "eosjs": "^21.1.0-29-d765fc3",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0"
+  }
+}
