--- webapp/nginx.conf
@@ -1,20 +1,0 @@
-events {}
-
-http {
-    server {
-        listen 8000 default_server;
-
-        server_name _;
-
-        location / {
-            proxy_pass http://127.0.0.1:8080;
-            proxy_http_version 1.1;
-            proxy_set_header Upgrade $http_upgrade;
-            proxy_set_header Connection "Upgrade";
-        }
-
-        location /v1/ {
-            proxy_pass http://127.0.0.1:8888;
-        }
-    }
-}
