--- webapp/package.json
@@ -1,27 +1,0 @@
-{
-  "private": true,
-  "description": "Example Website",
-  "main": "index.html",
-  "scripts": {
-    "serve": "webpack-dev-server --mode development"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/react": "^16.9.44",
-    "@types/react-dom": "^16.9.8",
-    "css-loader": "^4.2.0",
-    "source-map-loader": "^1.0.1",
-    "style-loader": "^1.2.1",
-    "ts-loader": "^8.0.2",
-    "typescript": "^3.9.7",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
-  },
-  "dependencies": {
-    "eosjs": "21.1.0-20-c31c737",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
-  }
-}
